board and pieces show up, works in landscape as well

+++ app/src/main/java/edu/msu/blaida/project1/Bishop.java
@@ -8,7 +8,7 @@ import android.graphics.BitmapFactory;
  */
 
 public class Bishop extends Piece {
-    public Bishop(Context context, int x, int y, int player) {
+    public Bishop(Context context, float x, float y, int player) {
         super(context, x, y, player);
         if(player == 1){
             this.setBitmap(BitmapFactory.decodeResource(context.getResources(),R.drawable.chess_blt45));
