+++ app/src/main/java/edu/msu/blaida/project1/King.java
@@ -9,7 +9,7 @@ import android.graphics.BitmapFactory;
 
 public class King extends Piece {
 
-    public King(Context context, int x, int y, int player) {
+    public King(Context context, float x, float y, int player) {
         super(context, x, y, player);
         if(player == 1){
             this.setBitmap(BitmapFactory.decodeResource(context.getResources(),R.drawable.chess_klt45));
