+++ app/src/main/java/edu/msu/blaida/project1/Pawn.java
@@ -16,7 +16,7 @@ public class Pawn extends Piece {
      */
     private boolean firstMove = true;
 
-    public Pawn(Context context, int x, int y, int player){
+    public Pawn(Context context, float x, float y, int player){
         super(context, x, y, player);
         if(player == 1){
             this.setBitmap(BitmapFactory.decodeResource(context.getResources(),R.drawable.chess_plt45));
